AI-Outbound-BE-main/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
pydantic==2.11.2
pymongo==4.11.3
dnspython==2.7.0
motor==3.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
retell-sdk
requests
schedule==1.2.1
email-validator==2.1.0.post1
bcrypt==4.0.1
httpx==0.28.1
msgraph-core==0.2.2
cloudinary==1.44.0